requirement.txt
alabaster==0.7.16
anyio==4.9.0
babel==2.17.0
boto3==1.37.32
botocore==1.37.32
certifi==2025.1.31
cfgv==3.4.0
chardet==3.0.4
charset-normalizer==3.4.1
colorama==0.4.6
distlib==0.3.9
dnspython==2.7.0
docutils==0.16
envs==1.4
filelock==3.18.0
futures==3.0.5
googletrans==3.1.0a0
greenlet==3.1.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
identify==2.6.9
idna==2.10
imagesize==1.4.1
Jinja2==3.1.6
jmespath==1.0.1
legacy-cgi==2.6.3
MarkupSafe==3.0.2
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.7
pre_commit==4.2.0
psycopg2==2.9.10
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rfc3986==1.5.0
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==5.3.0
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.40
typing_extensions==4.13.1
urllib3==2.3.0
virtualenv==20.30.0
xmltodict==0.14.2
